Move ExtractProtectedHeaders function to the crypto package
This commit moves the ExtractProtectedHeaders to the crypto package where it located together with other JWS/JWT related methods.
