+++ crypto/jwx_test.go
@@ -714,3 +714,117 @@ func Test_signingAlg(t *testing.T) {
 		assert.EqualError(t, err, "could not determine signature algorithm for key type '<nil>'")
 	})
 }
+
+func TestExtractProtectedHeaders(t *testing.T) {
+
+	var normalJws = func(claims map[string]interface{}) (string, error) {
+		jwk, err := GenerateJWK()
+		if err != nil {
+			return "", err
+		}
+		marshal, err := json.Marshal(claims)
+		if err != nil {
+			return "", err
+		}
+		sign, err := jws.Sign(marshal, jws.WithKey(jwa.ES256, jwk))
+		if err != nil {
+			return "", err
+		}
+		return string(sign), err
+	}
+	var doubleSignedJws = func(claims map[string]interface{}) (string, error) {
+		jwk, err := GenerateJWK()
+		if err != nil {
+			return "", err
+		}
+		marshal, err := json.Marshal(claims)
+		if err != nil {
+			return "", err
+		}
+		sign, err := jws.Sign(marshal, jws.WithKey(jwa.ES256, jwk), jws.WithKey(jwa.ES256, jwk), jws.WithJSON())
+		if err != nil {
+			return "", err
+		}
+		return string(sign), err
+	}
+	var noSignedJws = func(claims map[string]interface{}) (string, error) {
+		marshal, err := json.Marshal(claims)
+		if err != nil {
+			return "", err
+		}
+		sign, err := jws.Sign(marshal, jws.WithInsecureNoSignature())
+		if err != nil {
+			return "", err
+		}
+		return string(sign), err
+	}
+
+	jwt, err := normalJws(map[string]interface{}{"iss": "test"})
+	if err != nil {
+		t.Error(err)
+	}
+	double, err := doubleSignedJws(map[string]interface{}{"iss": "test"})
+	if err != nil {
+		t.Error(err)
+	}
+	none, err := noSignedJws(map[string]interface{}{"iss": "test"})
+	if err != nil {
+		t.Error(err)
+	}
+	testCases := []struct {
+		name          string
+		jwt           string
+		expectResults bool
+		expectError   error
+	}{
+		{
+			name:          "ValidJWT",
+			jwt:           jwt,
+			expectResults: true,
+		},
+		{
+			name:          "too many signatures",
+			jwt:           double,
+			expectResults: false,
+			expectError:   ErrorInvalidNumberOfSignatures,
+		},
+		{
+			name:          "no signatures",
+			jwt:           none,
+			expectResults: true,
+		},
+		{
+			name: "InvalidJWTHeader",
+			jwt:  "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiIsIng1YyI6dHJ1ZX0.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.fyenaNFjX705H02aOrpHayRVHa1uVxpQRUxWCl91rB4",
+		},
+		{
+			name: "InvalidJWT",
+			jwt:  "invalidToken",
+		},
+		{
+			name: "EmptyJWT",
+			jwt:  "",
+		},
+	}
+
+	for _, tc := range testCases {
+		t.Run(tc.name, func(t *testing.T) {
+			headers, err := ExtractProtectedHeaders(tc.jwt)
+			if err != nil {
+				if tc.expectError == nil {
+					t.Errorf("ExtractProtectedHeaders() error = %v", err)
+				} else if err.Error() != tc.expectError.Error() {
+					t.Errorf("ExtractProtectedHeaders() error = %v, expected: %v", err, tc.expectError)
+				}
+			} else {
+				if !tc.expectResults && len(headers) > 0 {
+					t.Errorf("ExtractProtectedHeaders() = %v, expected an empty header map", headers)
+				} else if tc.expectResults {
+					if _, ok := headers["alg"]; ok == false {
+						t.Errorf("ExtractProtectedHeaders() = %v, expected a valid header map", headers)
+					}
+				}
+			}
+		})
+	}
+}
